requirements.txt
anyio==4.9.0
certifi==2025.4.26
contourpy==1.3.2
CTkMessagebox==2.7
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
exceptiongroup==1.3.0
flet==0.28.2
flet-desktop==0.28.2
fonttools==4.58.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
mpmath==1.3.0
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
repath==0.9.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
typing_extensions==4.13.2
